Refactor to move the utility functions about setting up application by server gateway to meaning place.
