+++ test/unit_test/server/init.py
@@ -1,22 +1,15 @@
-import re
-from typing import Callable, Type, Union
+from typing import Type
 from unittest.mock import Mock, patch
 
 import pytest
 
 import pymock_api.server as mock_server
-from pymock_api.exceptions import FunctionNotFoundError
 
 from ..._values import _Test_Config
 
 mock_flask_server = Mock(mock_server.FlaskServer())
 mock_fastapi_server = Mock(mock_server.FastAPIServer())
 mock_server_obj = Mock(mock_server.MockHTTPServer)
-
-
-def _fake_function() -> None:
-    # This is fake function for PyTest
-    pass
 
 
 class TestLoadApp:
@@ -85,39 +78,3 @@ class TestCreateAppFunctions:
     def test_create_flask_app(self, mock_load_app: Mock):
         mock_server.create_fastapi_app()
         mock_load_app.assert_called_once()
-
-
-class TestSetupServerGateway:
-    @pytest.mark.parametrize(
-        ("app", "expected_app"),
-        [("create_flask_app()", "create_flask_app()"), (mock_server.create_flask_app, "create_flask_app()")],
-    )
-    @patch("pymock_api.server.WSGIServer")
-    def test_setup_wsgi(self, mock_instantiate_wsgi: Mock, app: Union[str, Callable], expected_app: str):
-        mock_server.setup_server_gateway.wsgi(web_app=app)
-        mock_instantiate_wsgi.assert_called_once_with(app=expected_app)
-
-    @pytest.mark.parametrize(
-        ("sut_func", "app", "expected_error"),
-        [
-            ("wsgi", "create_not_exist_app()", FunctionNotFoundError),
-            ("wsgi", _fake_function, FunctionNotFoundError),
-        ],
-    )
-    @patch("pymock_api.server.WSGIServer")
-    def test_bad_setup_wsgi(
-        self, mock_instantiate_wsgi: Mock, sut_func: str, app: Union[str, Callable], expected_error: Exception
-    ):
-        with pytest.raises(expected_error) as exc_info:
-            getattr(mock_server.setup_server_gateway, sut_func)(web_app=app)
-        assert re.search(r"cannot find .{0,32}function", str(exc_info.value), re.IGNORECASE)
-        mock_instantiate_wsgi.assert_not_called()
-
-    @pytest.mark.parametrize(
-        ("app", "expected_app"),
-        [("create_fastapi_app", "create_fastapi_app"), (mock_server.create_fastapi_app, "create_fastapi_app")],
-    )
-    @patch("pymock_api.server.ASGIServer")
-    def test_setup_asgi(self, mock_instantiate_asgi: Mock, app: Union[str, Callable], expected_app: str):
-        mock_server.setup_server_gateway.asgi(web_app=app)
-        mock_instantiate_asgi.assert_called_once_with(app=expected_app)
